Refactor Combat class and update method signatures: Moved player reference in Combat.h, adjusted HP variable positioning, and modified parameter name in executeExplore method for clarity in ExploreCounter.

+++ src/UI/Terminal.cpp
@@ -153,7 +153,7 @@ std::string Terminal::GetLine(int maxLength) {
                 lKeyCallback();
             }
         }
-        if (input.length() < maxLength) {
+        if (input.length() < static_cast<size_t>(maxLength)) {
             input += ch;
         }
     }
